+++ packages/pluggableWidgets/background-gradient-native/src/__tests__/backgroundGradient.spec.tsx
@@ -2,27 +2,55 @@ import { createElement } from "react";
 import { Text } from "react-native";
 import { render, fireEvent } from "@testing-library/react-native";
 import Big from "big.js";
-
 import { ActionValue } from "mendix";
-
 import { BackgroundGradient, props } from "../BackgroundGradient";
 
 describe("Background gradient", () => {
     let defaultProps: props;
     beforeEach(() => {
         defaultProps = {
-            angle: 0,
             colorList: [
                 { color: "#fff", offset: Big(0) },
                 { color: "#000", offset: Big(1) }
             ],
             content: <Text>Test</Text>,
             name: "test",
-            opacity: Big(1),
             style: []
         };
     });
     it("render background gradient correctly", () => {
+        const component = render(<BackgroundGradient {...defaultProps} />);
+        expect(component.toJSON()).toMatchSnapshot();
+    });
+    it("should throw error when colorList is empty", () => {
+        defaultProps = {
+            ...defaultProps,
+            colorList: []
+        };
+        expect(() => render(<BackgroundGradient {...defaultProps} />)).toThrowError();
+    });
+    it("render background gradient with custom style", () => {
+        defaultProps = {
+            ...defaultProps,
+            colorList: [],
+            style: [
+                {
+                    container: {},
+                    colorList: [
+                        { color: "#fff", offset: Big(0) },
+                        { color: "#000", offset: Big(1) }
+                    ]
+                }
+            ]
+        };
+        const component = render(<BackgroundGradient {...defaultProps} />);
+        expect(component.toJSON()).toMatchSnapshot();
+    });
+    it("render background gradient with one color", () => {
+        defaultProps = {
+            ...defaultProps,
+            colorList: [{ color: "#fff", offset: Big(0) }]
+        };
         const component = render(<BackgroundGradient {...defaultProps} />);
         expect(component.toJSON()).toMatchSnapshot();
     });