+++ packages/pluggableWidgets/background-gradient-native/src/ui/Styles.ts
@@ -1,12 +1,17 @@
 import { ViewStyle } from "react-native";
-import { Style } from "@mendix/pluggable-widgets-tools";
+import { ColorListType } from "../../typings/BackgroundGradientProps";
 
-export interface CustomStyle extends Style {
+export interface CustomStyle {
     container: ViewStyle;
+    angle?: number;
+    colorList?: ColorListType[];
+    opacity?: number;
 }
 
 const defaultStyle: CustomStyle = {
-    container: {}
+    container: {},
+    opacity: 1,
+    angle: 0
 };
 
 export default defaultStyle;