+++ packages/jsActions/nanoflow-actions-native/typings/mx.d.ts
@@ -1,9 +1,15 @@
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 declare namespace mx {
-    interface ui extends mx.ui {
+    interface ui {
         toggleSidebar: () => void;
     }
-    interface data extends mx.data {
+    interface data {
         update: (param: { guid?: string | undefined; entity?: string | undefined; callback?: () => void }) => void;
+    }
+    interface session {
+        clearCachedSessionData: () => Promise<void>;
+    }
+    interface MxInterface {
+        reload: () => void;
     }
 }