+++ packages/theming/atlas/src/themesource/atlas_core/native/types/variables.d.ts
@@ -3,7 +3,7 @@
 ========================================================================== */
 
 import { ViewStyle } from "react-native";
-import { CheckBoxInputType } from "./widgets";
+import { CheckBoxInputType, GradientColorList } from "./widgets";
 
 declare type FontWeight = "normal" | "bold" | "100" | "200" | "300" | "400" | "500" | "600" | "700" | "800" | "900";
 declare type TextAlign = "auto" | "left" | "right" | "center" | "justify";
@@ -589,6 +589,9 @@ export interface VariablesRadioButtons {
 
 export interface VariableBackgroundGradient {
     container: ViewStyle;
+    angle: number;
+    colorList?: GradientColorList[];
+    opacity: number;
 }
 
 interface ColumnChartGridStyle {